% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertGoogleApiToSmoof.R
\name{convertGoogleApiToSmoof}
\alias{convertGoogleApiToSmoof}
\title{Converts a Google Maps API Call to Single Objective Function}
\usage{
convertGoogleApiToSmoof(lat.range, lon.range, maximize = TRUE,
  google.apikey = NULL)
}
\arguments{
\item{lat.range}{[\code{numeric(2)}]}

\item{lon.range}{[\code{numeric(2)}]}

\item{maximize}{[\code{logic(2)}]}

\item{google.apikey}{[\code{character(1)}]
The Google Maps Elevation API Key.
Will be read from config file.}
}
\value{
\code{smoof function}
}
\description{
Receives the JSON from Google Maps API and extracts height
}
