% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveGeo2SmoofConfig.R
\name{saveGeo2SmoofConfig}
\alias{saveGeo2SmoofConfig}
\title{Saves a list of geo2smoof configuration settings to file.}
\usage{
saveGeo2SmoofConfig(google.apikey = NULL, overwrite = FALSE)
}
\arguments{
\item{google.apikey}{[\code{character(1)}]
The Google Maps Elevation API Key.}

\item{overwrite}{[\code{logical(1)}]\cr
Should an existing file be overwritten?
Default is \code{FALSE}.}
}
\description{
The new configuration is automatically assigned.
}
